requirements.txt
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-extensions==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
model-bakery==1.3.2
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.0
wcwidth==0.2.5
whitenoise==5.3.0
